+++ lib/src/task.dart
@@ -28,11 +28,9 @@ class AlgoliaTask {
   }
 
   Future<bool> taskStatus() async {
-    var url = '${algolia._host}indexes/$_index/task/$taskID';
-
-    var response = await http.get(
-      Uri.parse(url),
-      headers: algolia._headers,
+    var response = await algolia._apiCall(
+      ApiRequestType.get,
+      'indexes/$_index/task/$taskID',
     );
     Map<String, dynamic> body = json.decode(response.body);
     if (!(response.statusCode >= 200 && response.statusCode < 300)) {