+++ CHANGELOG.md
@@ -1,8 +1,35 @@
-## [1.0.3] - Bug fixes
+# Changelog
+## 1.0.4
+### Implemented Fallback Request
+
+- [Added] You can now worry less about retry action action on failure, as it just do it for you. In order to guarantee a very high availability, we implemented recommended retry strategy for all API calls on all your read and write actions. (Currently, fallback request is not supported for insight.)
+- [Added] `AlgoliaSynonymsReference` now easily set synonyms with just few lines of code
+  ```dart
+    // single
+    algolia.index('contacts').synonyms.save(AlgoliaSynonyms(
+      objectID: '1',
+      type: SynonymsType.synonym,
+      synonyms: ['iphne', 'iphone', 'ipone'],
+      forwardToReplicas: true,
+    ));
+    
+    // batch
+    algolia.index('contacts').synonyms.batch([
+      AlgoliaSynonyms(
+        objectID: '1',
+        type: SynonymsType.synonym,
+        synonyms: ['iphne', 'iphone', 'ipone'],
+        forwardToReplicas: true,
+      ),
+    ]);
+  ```
+## 1.0.3
+### Bug fixes
 
 - [Bug] queryId from QuerySnapshot was made nullable.
 
-## [1.0.2] - Added support for Facet Values & Insights
+## 1.0.2
+### Added support for Facet Values & Insights
 
 - [Added] Facet values `AlgoliaFacetValueSnapshot`:
   Now you can get list of all facet value to implement advance filtering options.
@@ -25,7 +52,8 @@
 - [Improved] Improved concurrency of snapshot interface by making constructor base multiple mapped value to a getter parameters.
 
 
-## [1.0.1] - Bug-fixes with improved debugging stability
+## 1.0.1
+### Bug-fixes with improved debugging stability
 
 - [Bug] faulty assert resolved for checking empty values [#40](https://github.com/knoxpo/dart_algolia/issues/40)
 - [Bug] Fixed all enum valued query and setting methods.
@@ -39,11 +67,13 @@
 - [Added] Improved stability for debugging use ``.toString()`` to get working variables of the interface (applicable for all Algolia classes).
 - [Added] ``.toMap()`` to all data dictionary classes.
 
-## [1.0.0+1] - Stable release with Null-Safety
+## 1.0.0+1
+### Stable release with Null-Safety
 
 - Implemented ``analysis_options.yaml``
 
-## [1.0.0] - Stable release with Null-Safety
+## 1.0.0
+### Stable release with Null-Safety
 
 - [Bug] [#26](https://github.com/knoxpo/dart_algolia/issues/26)
 - [Added] Add support of ``Null-safety``
@@ -65,16 +95,19 @@
 - [Added] Add new error handling class ``AlgoliaError``
 - [Upgrade] Bumped up ``http`` version
 
-## [0.1.7] - Bug fixes and added a new property
+## 0.1.7
+### Bug fixes and added a new property
 
 - [Bug] [#14](https://github.com/knoxpo/dart_algolia/issues/14) Solved few health suggestion, to improve the health of the code.
 - [Added] Add support for ``facets_stats`` property returned by Algolia query
 
-## [0.1.6+1] - Improve library health
+## 0.1.6+1
+### Improve library health
 
 - [Bug] Solved few health suggestion, to improve the health of the code.
 
-## [0.1.6] - Added Multi-Query
+## 0.1.6
+### Added Multi-Query
 
 - [Added] PR implementation of ``multipleQueries``
 
@@ -94,17 +127,20 @@
         await algolia.multipleQueries.addQueries([queryA, queryB]).getObjects();  
   ```
 
-## [0.1.5] - Added New Functionalities
+## 0.1.5
+### Added New Functionalities
 
 - [Bug] Solved a technical reported bug [#11](https://github.com/knoxpo/dart_algolia/issues/11)
 - [Added] Copy, Move Index functionalities.
 - [Added] PR implementation of ``replaceAllObjects()``
 
-## [0.1.4+3] - Improve library health
+## 0.1.4+3
+### Improve library health
 
 - [Bug] Solved few health suggestion, to improve the health of the code.
 
-## [0.1.4+2] - Added few advance query references and solved bugs
+## 0.1.4+2
+### Added few advance query references and solved bugs
 
 - [Bug] `.setFacetFilter(dynamic value)` can now accept String or List<String> value.
 - [Added] AttributeForDistinct (Advance)
@@ -112,25 +148,30 @@
 - [Added] GetRankingInfo (Advance)
 - [Added] ClickAnalytics (Advance)
 
-## [0.1.4+1] - Added support facets
+## 0.1.4+1
+### Added support facets
 
 - Added `facets` to ``AlgoliaQuerySnapshot`` to list facets name with hits count.
 
-## [0.1.3+2] - Implementation & bug solved
+## 0.1.3+2
+### Implementation & bug solved
 
 - highlightResult [Bug] (commit ref: 0d76d24fe8aa347a0933920afe5ded43bdcbd68b)
 - snippetResult [Implementation] (commit ref: 0d76d24fe8aa347a0933920afe5ded43bdcbd68b)
 
-## [0.1.3+1] - Added support to manage index settings
+## 0.1.3+1
+### Added support to manage index settings
 
 - Updated `example.dart`: Added index settings example.
 - Updated index `.setSettings()` response to `AlgoliaTask`.
 
-## [0.1.3] - Added support to manage index settings
+## 0.1.3
+### Added support to manage index settings
 
 - Added support to manage index settings (Get & Set), limited to 24 settings parameters, more to be added in newer releases.
 
-## [0.1.2] - Added new query params
+## 0.1.2
+### Added new query params
 
 - OptionalFilter (Filtering)
 - NumericFilter (Filtering)
@@ -151,11 +192,13 @@
 - DisableTypoToleranceOnWords (Typo)
 - SeparatorsToIndex (Typo)
 
-## [0.1.1] - Added example
+## 0.1.1
+### Added example
 
 - Bug fixes.
 - Removed Flutter direct dependency to support universal dart projects.
 
-## [0.1.0] - Initial Release
+## 0.1.0
+### Initial Release
 
 - Initial release.